+++ .all_custom_gradle_tasks/Android_App.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Android_App" type="AndroidRunConfigurationType" factoryName="Android App" singleton="true">
-    <module name="ComposeMultiplatformTemplate.composeApp.main" />
+    <module name="Reluct.composeApp.main" />
     <option name="DEPLOY" value="true" />
     <option name="DEPLOY_APK_FROM_BUNDLE" value="false" />
     <option name="DEPLOY_AS_INSTANT" value="false" />
