+++ common/src/androidMain/kotlin/work/racka/reluct/common/sources/DbClient.android.kt
@@ -1,0 +1,12 @@
+package work.racka.reluct.common.sources
+
+import android.content.Context
+import android.os.Build
+
+internal actual class DbClient(
+    private val context: Context
+) {
+    actual fun platformName(): String {
+        return "Android ${Build.VERSION.SDK_INT}, pkg: ${context.packageName}"
+    }
+}
