+++ common/src/commonMain/kotlin/work/racka/reluct/common/di/KoinMain.kt
@@ -1,0 +1,17 @@
+package work.racka.reluct.common.di
+
+import org.koin.core.context.startKoin
+import org.koin.dsl.KoinAppDeclaration
+
+object KoinMain {
+    fun init(
+        appDeclaration: KoinAppDeclaration = {}
+    ) {
+        startKoin {
+            appDeclaration()
+
+            // Initialize modules
+            Common.run { install() }
+        }
+    }
+}
