+++ common/src/commonMain/kotlin/work/racka/reluct/common/sources/DbClient.kt
@@ -1,0 +1,5 @@
+package work.racka.reluct.common.sources
+
+internal expect class DbClient {
+    fun platformName(): String
+}
