+++ common/src/commonMain/kotlin/work/racka/reluct/common/sources/MyRepository.kt
@@ -1,0 +1,13 @@
+package work.racka.reluct.common.sources
+
+interface MyRepository {
+    fun helloWorld(): String
+}
+
+internal class MyRepositoryImpl(
+    private val dbClient: DbClient
+) : MyRepository {
+    override fun helloWorld(): String {
+        return "Hello: ${dbClient.platformName()}"
+    }
+}
