+++ common/src/commonMain/kotlin/work/racka/reluct/common/sources/MyViewModel.kt
@@ -1,0 +1,10 @@
+package work.racka.reluct.common.sources
+
+import androidx.lifecycle.ViewModel
+
+class MyViewModel(
+    private val repository: MyRepository
+) : ViewModel() {
+
+    fun getHelloWorldString(): String = repository.helloWorld()
+}
