--- common/src/commonMain/kotlin/work/racka/template/common/di/KoinMain.kt
@@ -1,17 +1,0 @@
-package work.racka.template.common.di
-
-import org.koin.core.context.startKoin
-import org.koin.dsl.KoinAppDeclaration
-
-object KoinMain {
-    fun init(
-        appDeclaration: KoinAppDeclaration = {}
-    ) {
-        startKoin {
-            appDeclaration()
-
-            // Initialize modules
-            Common.run { install() }
-        }
-    }
-}
