--- common/src/commonMain/kotlin/work/racka/template/common/sources/MyViewModel.kt
@@ -1,10 +1,0 @@
-package work.racka.template.common.sources
-
-import androidx.lifecycle.ViewModel
-
-class MyViewModel(
-    private val repository: MyRepository
-) : ViewModel() {
-
-    fun getHelloWorldString(): String = repository.helloWorld()
-}
