+++ common/src/desktopMain/kotlin/work/racka/reluct/common/di/Common.desktop.kt
@@ -1,0 +1,10 @@
+package work.racka.reluct.common.di
+
+import org.koin.core.module.Module
+import org.koin.core.module.dsl.singleOf
+import org.koin.dsl.module
+import work.racka.reluct.common.sources.DbClient
+
+internal actual fun platform(): Module = module {
+    singleOf(::DbClient)
+}
