+++ composeApp/src/androidMain/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">KMP Hello World</string>
+    <string name="app_name">Reluct</string>
 </resources>
