+++ tooling/plugins/src/main/kotlin/work/racka/reluct/extensions/Versions.kt
@@ -1,0 +1,12 @@
+package work.racka.reluct.extensions
+
+internal object Versions {
+    const val COMPILE_SDK = 34
+    val MIN_SDK = 26
+    val TARGET_SDK = 34
+    val VERSION_CODE = 1
+    val VERSION_NAME = "1.0.0"
+    val PACKAGE_NAME = "work.racka.reluct"
+    val DESKTOP_VERSION = "1.0.0"
+    val DESKTOP_PACKAGE_NAME = "Reluct"
+}
