--- tooling/plugins/src/main/kotlin/work/racka/template/extensions/Versions.kt
@@ -1,12 +1,0 @@
-package work.racka.template.extensions
-
-internal object Versions {
-    const val COMPILE_SDK = 34
-    val MIN_SDK = 26
-    val TARGET_SDK = 34
-    val VERSION_CODE = 1
-    val VERSION_NAME = "1.0.0"
-    val PACKAGE_NAME = "org.example.project"
-    val DESKTOP_VERSION = "1.0.0"
-    val DESKTOP_PACKAGE_NAME = "template"
-}
