--- tooling/plugins/src/main/kotlin/work/racka/template/extensions/VersionsCatalog.kt
@@ -1,9 +1,0 @@
-package work.racka.template.extensions
-
-import org.gradle.api.Project
-import org.gradle.api.artifacts.VersionCatalog
-import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.kotlin.dsl.getByType
-
-internal val Project.libs: VersionCatalog
-    get() = extensions.getByType<VersionCatalogsExtension>().named("libs")
